+++ CardEngine/src/TurnSystem/localPlayer.h
@@ -16,7 +16,7 @@ namespace turnSystem
         std::shared_ptr<eventBus::eventBus> events;
 
     public:
-        localPlayer(std::string name, std::shared_ptr<eventBus::eventBus> events, size_t id) : IPlayer(id, name),
+        localPlayer(std::string name, std::shared_ptr<eventBus::eventBus> events, uint16_t id) : IPlayer(id, name),
             events(events)
         {
         }
